Fix error when invoking magit-status on Window
I had ripgrep in my PATH on Windows and noticed it started causing
error when invoking magit-status. This is because nice is not really
available. Instead of trying to be clever and looking up if nice
exists in the PATH (which is unlikely/impossible? on Windows) I have
simply disabled the use of nice for magit-todos if on Windows. In my
opinion, this should probably be the default for the magit-todos
package but meh, better be eplicit :-)